refactor: Refactor to separate view and logic of image field
